+++ packages/ricos-editor/web/src/modals/EditorModalProvider.tsx
@@ -1,7 +1,7 @@
 import React, { Children, Component, ReactElement, Suspense, Fragment } from 'react';
 import mergeModalStyles from './mergeModalStyles';
 import { ModalStyles } from 'wix-rich-content-common';
-import { ModalsMap, ModalSettings, RichContentProps } from '../index';
+import { ModalsMap, ModalSettings } from '../index';
 import { merge } from 'lodash';
 
 interface Props {
@@ -29,7 +29,7 @@ interface State {
 }
 
 export default class EditorModalProvider extends Component<Props, State> {
-  childProps: RichContentProps;
+  modalHandlers: { helpers: ModalSettings };
 
   constructor(props: Props) {
     super(props);
@@ -37,10 +37,8 @@ export default class EditorModalProvider extends Component<Props, State> {
       showModal: false,
       editorModalId: `EditorModal-${new Date().getTime()}`,
     };
-    this.childProps = {
-      ...props.children.props,
+    this.modalHandlers = {
       helpers: {
-        ...props.children.props.helpers,
         openModal: this.openModal,
         closeModal: this.closeModal,
       },
@@ -82,9 +80,10 @@ export default class EditorModalProvider extends Component<Props, State> {
   render() {
     const { EditorModal, showModal, modalProps, modalStyles, editorModalId } = this.state;
     const { children, ModalsMap, locale, theme, ariaHiddenId } = this.props;
+    const childProps = merge(children.props, this.modalHandlers);
     return (
       <Fragment>
-        {Children.only(React.cloneElement(children, { ...this.childProps }))}
+        {Children.only(React.cloneElement(children, childProps))}
         <div>
           <div id={editorModalId} />
           {EditorModal && (