+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     Click to see more.
   </summary>
 </details>
+
+### :bug: Bug Fix
+- `ricos-editor`
+  - [#1508](https://github.com/wix-incubator/rich-content/pull/1508) child (RCE) props were never updated
+
 <hr/>
 
 ## 7.18.2 (Sep 3, 2020)